Set kind of SootClass only if tracked by Scene
This ensures that the `classes` field of `Scene` will always contain
a superset of the application, library, and phantom classes. This
invariant seem to be implied but not properly enforced by `SootClass`.

The setter methods of `SootClass` that mark it as a specific kind (e.g.,
`setApplicationclass()`), may add it to the relevant subset (e.g.,
`applicationClasses`), even if it hasn't been added to the set of all
classes yet.